refactor(app): extract prompt-input derived state into derived-state.ts
Move the 12 read-only derived memos (info/status/working/imageAttachments/
actionReady/abortReady/commentCount/blank/stopping/contextItems/placeholder/
accepting) out of the PromptInput component into createPromptDerivedState.
Pure extraction, no behavior change.
